Remove exporter.exporter Exporter, no longer updating DSS
